job-outreach-pipeline/requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.23.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
